refactor: extract snake board component
